+++ R/index.R
@@ -1,4 +1,6 @@
 
+`_renv_index` <- new.env(parent = emptyenv())
+
 index <- function(scope, key = NULL, callback = NULL, limit = 3600L) {
 
   enabled <- renv_index_enabled(scope, key)
@@ -64,11 +66,28 @@ renv_index_get <- function(root, scope, index, key, now, limit) {
   if (renv_index_expired(entry, now, limit))
     return(NULL)
 
+  # check for in-memory cached value
+  value <- `_renv_index`[[scope]][[key]]
+  if (!is.null(value))
+    return(value)
+
+  # otherwise, try to read from disk
   data <- file.path(root, entry$data)
   if (!file.exists(data))
     return(NULL)
 
-  readRDS(data)
+  # read data from disk
+  value <- readRDS(data)
+
+  # add to in-memory cache
+  `_renv_index`[[scope]] <-
+    `_renv_index`[[scope]] %||%
+    new.env(parent = emptyenv())
+
+  `_renv_index`[[scope]][[key]] <- value
+
+  # return value
+  value
 
 }
 
@@ -83,7 +102,7 @@ renv_index_set <- function(root, scope, index, key, callback, now, limit) {
   saveRDS(value, file = data, version = 2L)
 
   # clean up stale entries
-  index <- renv_index_clean(root, index, now, limit)
+  index <- renv_index_clean(root, scope, index, now, limit)
 
   # add index entry
   index[[key]] <- list(time = now, data = basename(data))
@@ -103,17 +122,39 @@ renv_index_encode <- function(key) {
   renv_hash_text(text)
 }
 
-renv_index_clean <- function(root, index, now, limit) {
+renv_index_clean <- function(root, scope, index, now, limit) {
 
   # figure out what cache entries have expired
-  expired <- map_lgl(index, renv_index_expired, now = now, limit = limit)
-
-  # remove the expired cache entries
-  paths <- map_chr(index, `[[`, "data")
-  unlink(file.path(root, paths))
+  ok <- enum_lgl(
+    index,
+    renv_index_clean_impl,
+    root  = root,
+    scope = scope,
+    index = index,
+    now   = now,
+    limit = limit
+  )
 
   # return the existing cache entries
-  index[!expired]
+  index[ok]
+
+}
+
+renv_index_clean_impl <- function(key, entry, root, scope, index, now, limit) {
+
+  # check if cache entry has expired
+  expired <- renv_index_expired(entry, now, limit)
+  if (!expired)
+    return(TRUE)
+
+  # remove from in-memory cache
+  cache <- `_renv_index`[[scope]]
+  cache[[key]] <- NULL
+
+  # remove from disk
+  unlink(file.path(root, entry$data))
+
+  FALSE
 
 }